connectivity-exporter: monitoring all interfaces
The network interface name could be specified with the "-i" CLI flag but
it was not possible to monitor all network interfaces.

With this patch, connectivity-exporter will monitor all network
interfaces when the "-i" flag is empty or missing.

It works by setting sll_ifindex to zero: see "man 7 packet":

> sll_ifindex  is the interface index of the interface (see
> netdevice(7)); 0 matches any interface (only permitted for binding).