Always delete the connection from the BPF map after it has been counted
Previously this was done in the else statement which could lead to leaking
connections in the map. This happens whenever there are multiple
connections with equal SNIs in the same second. A leaked connection entry
will exist for 20s until the go code cleans up the map. After this change
there will be no more leaked entries.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Alban Crequy <albancrequy@microsoft.com>